KKL final

+++ database/migrations/2025_09_26_223443_create_items_table.php
@@ -6,9 +6,6 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::create('items', function (Blueprint $table) {
@@ -16,15 +13,12 @@ return new class extends Migration
             $table->string('code')->nullable();
             $table->string('category')->nullable();
             $table->string('name');
-            $table->decimal('quantity', 10, 2)->default(0);
+            $table->unsignedInteger('quantity')->default(0);
             $table->string('status')->default('available');
             $table->timestamps();
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('items');
